Reduced 64-bit build warnings. Removed UNDER_CE. Remove Windows version check.

+++ Library/TeamTalkLib/codec/WaveFile.cpp
@@ -37,7 +37,7 @@ bool WriteWaveFileHeader(ACE_FILE_IO& file, const media::AudioFormat& fmt)
     waveformat.nChannels = fmt.channels;
     waveformat.nSamplesPerSec = fmt.samplerate;
     waveformat.nAvgBytesPerSec = PCM16_BYTES(fmt.samplerate, fmt.channels);
-    waveformat.nBlockAlign = PCM16_BYTES(1, fmt.channels);
+    waveformat.nBlockAlign = uint16_t(PCM16_BYTES(1, fmt.channels));
     waveformat.wBitsPerSample = 16;
     waveformat.cbSize = 0;
